Apply translate to Star
