+++ shoes-swt/lib/shoes/swt.rb
@@ -75,6 +75,7 @@ class Shoes
       require 'shoes/swt/oval_painter.rb'
       require 'shoes/swt/rect_painter.rb'
       require 'shoes/swt/shape_painter.rb'
+      require 'shoes/swt/star_painter.rb'
       require 'shoes/swt/swt_button'
       require 'shoes/swt/check_button'
       require 'shoes/swt/radio_group'
